+++ src/test/TestConnectionOptions2dx.h
@@ -9,8 +9,6 @@
 #ifndef ADJUST_TESTCONNECTIONOPTIONS2DX_H_
 #define ADJUST_TESTCONNECTIONOPTIONS2DX_H_
 
-#include <iostream>
-
 class TestConnectionOptions2dx {
 public:
     static void setTestConnectionOptions();